Move expressions to separate file
